requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
asyncpg==0.29.0
python-dotenv==1.0.1
pillow==10.2.0
pytesseract==0.3.10
APScheduler==3.10.4
google-generativeai==0.3.2
